requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
faker>=37.0.0
